Worked on base system.
Finished up the crawling section of this application by adding a lot of the
global code required to glue everything together. This includes actually
implementing the -t option, adding a -q quiet option for the echo log, and
making a setting for the root folder where all species data is stored
locally.

Also changed the directory structure so the crawler name is appended to the
subdirectories instead of being a root directory so everything is now sorted
by taxonomy species id.