Refactoring to process/task system.
After attempting a quick hacky fix last night it failed because of the
following; for every hack that was made in the code it caused two additional
problems in other code. Eventually it was realized that there is no easy fix so
refactoring began.

Made very good progress in a new process/task system. The abstract process
interface replaces the mirror interface. The new process interface adds
additional freedom to provide custom indexing tasks for each process type, the
core thing that is required for the necessary changes in NCBI. Mirroring within
the new process interface is now also fully customizable. The abstract task
interface provides the building blocks for customizing the mirroring and
indexing processes.

The old system is still present but deprecated. Many tasks still need to be
done and indexing has not been redone yet to use the new process/task system.