Fixed cdna task bug.
Resolved bug with write cdna task for NCBI database assemblies where the
temporary .fa.fai file and incomplete cdna file would not be deleted if the
gffread command failed. This in turn would cause pynome to run the kallisto and
salmon tasks, thinking it can because the cdna file existed within the assembly
directory.